gestion des token ok

+++ laravel-API/routes/api.php
@@ -10,25 +10,25 @@ Route::get('/v1/welcome', function () {
 });
 
 // User routes
-Route::get('v1/users', [UserController::class, 'index']);
-Route::post('v1/users', [UserController::class, 'store']);
-Route::get('v1/users/{user}', [UserController::class, 'show']);
-Route::put('v1/users/{user}', [UserController::class, 'update']);
-Route::delete('v1/users/{user}', [UserController::class, 'destroy']);
+Route::get('v1/users', [UserController::class, 'index'])->middleware('auth:sanctum');
+Route::post('v1/users', [UserController::class, 'store'])->middleware('auth:sanctum');
+Route::get('v1/users/{user}', [UserController::class, 'show'])->middleware('auth:sanctum');
+Route::put('v1/users/{user}', [UserController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('v1/users/{user}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
 
 // Product routes
-Route::get('v1/products', [ProductController::class, 'index']);
-Route::post('v1/products', [ProductController::class, 'store']);
-Route::get('v1/products/{product}', [ProductController::class, 'show']);
-Route::put('v1/products/{product}', [ProductController::class, 'update']);
-Route::delete('v1/products/{product}', [ProductController::class, 'destroy']);
+Route::get('v1/products', [ProductController::class, 'index'])->middleware('auth:sanctum');
+Route::post('v1/products', [ProductController::class, 'store'])->middleware('auth:sanctum');
+Route::get('v1/products/{product}', [ProductController::class, 'show'])->middleware('auth:sanctum');
+Route::put('v1/products/{product}', [ProductController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('v1/products/{product}', [ProductController::class, 'destroy'])->middleware('auth:sanctum');
 
 // Category routes
-Route::get('v1/categories', [CategoryController::class, 'index']);
-Route::post('v1/categories', [CategoryController::class, 'store']);
-Route::get('v1/categories/{category}', [CategoryController::class, 'show']);
-Route::put('v1/categories/{category}', [CategoryController::class, 'update']);
-Route::delete('v1/categories/{category}', [CategoryController::class, 'destroy']);
+Route::get('v1/categories', [CategoryController::class, 'index'])->middleware('auth:sanctum');
+Route::post('v1/categories', [CategoryController::class, 'store'])->middleware('auth:sanctum');
+Route::get('v1/categories/{category}', [CategoryController::class, 'show'])->middleware('auth:sanctum');
+Route::put('v1/categories/{category}', [CategoryController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('v1/categories/{category}', [CategoryController::class, 'destroy'])->middleware('auth:sanctum');
 
 Route::post('v1/login', [UserController::class, 'login']);
 Route::post('v1/register', [UserController::class, 'register']);
